lots of menu cleanup and some menu decomp
