+++ asm/kecleon_items.s
@@ -755,275 +755,4 @@ _080194F0: .4byte gKecleonShopDialogue
 _080194F4: .4byte 0x0000010d
 	thumb_func_end UpdateKecleonStoreDialogue
 
-	thumb_func_start sub_80194F8
-sub_80194F8:
-	push {r4-r7,lr}
-	ldr r4, _08019580
-	ldr r0, [r4]
-	adds r0, 0x74
-	movs r1, 0
-	movs r2, 0x10
-	bl MemoryFill16
-	ldr r3, [r4]
-	movs r2, 0
-	ldr r0, _08019584
-	ldr r0, [r0]
-	str r0, [r3, 0x34]
-	movs r0, 0x2
-	str r0, [r3, 0x38]
-	movs r1, 0x1
-	ldr r0, _08019588
-	ldr r0, [r0]
-	str r0, [r3, 0x3C]
-	movs r0, 0x3
-	str r0, [r3, 0x40]
-	ldr r0, _0801958C
-	str r0, [r3, 0x44]
-	movs r0, 0x4
-	str r0, [r3, 0x48]
-	ldr r0, _08019590
-	ldr r0, [r0]
-	str r0, [r3, 0x4C]
-	movs r0, 0x7
-	str r0, [r3, 0x50]
-	ldr r0, _08019594
-	ldr r0, [r0]
-	str r0, [r3, 0x54]
-	str r1, [r3, 0x58]
-	movs r6, 0x5
-	str r2, [r3, 0x5C]
-	str r1, [r3, 0x60]
-	movs r5, 0
-	mov r12, r3
-	adds r3, 0x38
-	mov r2, r12
-	adds r2, 0x74
-_0801954C:
-	ldrh r0, [r2]
-	cmp r0, 0
-	bne _0801955C
-	ldr r1, [r3]
-	mov r7, r12
-	ldr r0, [r7, 0x28]
-	cmp r1, r0
-	beq _080195BA
-_0801955C:
-	adds r3, 0x8
-	adds r2, 0x2
-	adds r5, 0x1
-	cmp r5, r6
-	blt _0801954C
-	movs r5, 0
-	cmp r5, r6
-	bge _080195BA
-	ldr r1, [r4]
-	adds r0, r1, 0
-	adds r0, 0x74
-	ldrh r0, [r0]
-	cmp r0, 0
-	bne _08019598
-	ldr r0, [r1, 0x38]
-	str r0, [r1, 0x28]
-	b _080195BA
-	.align 2, 0
-_08019580: .4byte gUnknown_203B210
-_08019584: .4byte gUnknown_80D4978
-_08019588: .4byte gUnknown_80D4984
-_0801958C: .4byte gUnknown_80DB8A0
-_08019590: .4byte gUnknown_80D4970
-_08019594: .4byte gUnknown_80D4934
-_08019598:
-	adds r5, 0x1
-	cmp r5, r6
-	bge _080195BA
-	ldr r2, [r4]
-	lsls r0, r5, 1
-	adds r1, r2, 0
-	adds r1, 0x74
-	adds r1, r0
-	ldrh r0, [r1]
-	cmp r0, 0
-	bne _08019598
-	lsls r0, r5, 3
-	adds r1, r2, 0
-	adds r1, 0x38
-	adds r1, r0
-	ldr r0, [r1]
-	str r0, [r2, 0x28]
-_080195BA:
-	pop {r4-r7}
-	pop {r0}
-	bx r0
-	thumb_func_end sub_80194F8
-
-	thumb_func_start sub_80195C0
-sub_80195C0:
-	push {r4-r7,lr}
-	ldr r4, _0801962C
-	ldr r0, [r4]
-	adds r0, 0x74
-	movs r1, 0
-	movs r2, 0x10
-	bl MemoryFill16
-	ldr r3, [r4]
-	movs r2, 0
-	ldr r0, _08019630
-	ldr r0, [r0]
-	str r0, [r3, 0x34]
-	movs r0, 0x2
-	str r0, [r3, 0x38]
-	movs r1, 0x1
-	ldr r0, _08019634
-	ldr r0, [r0]
-	str r0, [r3, 0x3C]
-	movs r0, 0x7
-	str r0, [r3, 0x40]
-	movs r6, 0x2
-	str r2, [r3, 0x44]
-	str r1, [r3, 0x48]
-	movs r5, 0
-	mov r12, r3
-	adds r3, 0x38
-	mov r2, r12
-	adds r2, 0x74
-_080195FA:
-	ldrh r0, [r2]
-	cmp r0, 0
-	bne _0801960A
-	ldr r1, [r3]
-	mov r7, r12
-	ldr r0, [r7, 0x30]
-	cmp r1, r0
-	beq _0801965A
-_0801960A:
-	adds r3, 0x8
-	adds r2, 0x2
-	adds r5, 0x1
-	cmp r5, r6
-	blt _080195FA
-	movs r5, 0
-	cmp r5, r6
-	bge _0801965A
-	ldr r1, [r4]
-	adds r0, r1, 0
-	adds r0, 0x74
-	ldrh r0, [r0]
-	cmp r0, 0
-	bne _08019638
-	ldr r0, [r1, 0x38]
-	str r0, [r1, 0x30]
-	b _0801965A
-	.align 2, 0
-_0801962C: .4byte gUnknown_203B210
-_08019630: .4byte gUnknown_80D4978
-_08019634: .4byte gUnknown_80D4970
-_08019638:
-	adds r5, 0x1
-	cmp r5, r6
-	bge _0801965A
-	ldr r2, [r4]
-	lsls r0, r5, 1
-	adds r1, r2, 0
-	adds r1, 0x74
-	adds r1, r0
-	ldrh r0, [r1]
-	cmp r0, 0
-	bne _08019638
-	lsls r0, r5, 3
-	adds r1, r2, 0
-	adds r1, 0x38
-	adds r1, r0
-	ldr r0, [r1]
-	str r0, [r2, 0x30]
-_0801965A:
-	pop {r4-r7}
-	pop {r0}
-	bx r0
-	thumb_func_end sub_80195C0
-
-	thumb_func_start sub_8019660
-sub_8019660:
-	push {r4-r7,lr}
-	ldr r4, _080196CC
-	ldr r0, [r4]
-	adds r0, 0x74
-	movs r1, 0
-	movs r2, 0x10
-	bl MemoryFill16
-	ldr r3, [r4]
-	movs r2, 0
-	ldr r0, _080196D0
-	ldr r0, [r0]
-	str r0, [r3, 0x34]
-	movs r0, 0x3
-	str r0, [r3, 0x38]
-	movs r1, 0x1
-	ldr r0, _080196D4
-	ldr r0, [r0]
-	str r0, [r3, 0x3C]
-	movs r0, 0x7
-	str r0, [r3, 0x40]
-	movs r6, 0x2
-	str r2, [r3, 0x44]
-	str r1, [r3, 0x48]
-	movs r5, 0
-	mov r12, r3
-	adds r3, 0x38
-	mov r2, r12
-	adds r2, 0x74
-_0801969A:
-	ldrh r0, [r2]
-	cmp r0, 0
-	bne _080196AA
-	ldr r1, [r3]
-	mov r7, r12
-	ldr r0, [r7, 0x2C]
-	cmp r1, r0
-	beq _080196FA
-_080196AA:
-	adds r3, 0x8
-	adds r2, 0x2
-	adds r5, 0x1
-	cmp r5, r6
-	blt _0801969A
-	movs r5, 0
-	cmp r5, r6
-	bge _080196FA
-	ldr r1, [r4]
-	adds r0, r1, 0
-	adds r0, 0x74
-	ldrh r0, [r0]
-	cmp r0, 0
-	bne _080196D8
-	ldr r0, [r1, 0x38]
-	str r0, [r1, 0x2C]
-	b _080196FA
-	.align 2, 0
-_080196CC: .4byte gUnknown_203B210
-_080196D0: .4byte gUnknown_80D4984
-_080196D4: .4byte gUnknown_80D4970
-_080196D8:
-	adds r5, 0x1
-	cmp r5, r6
-	bge _080196FA
-	ldr r2, [r4]
-	lsls r0, r5, 1
-	adds r1, r2, 0
-	adds r1, 0x74
-	adds r1, r0
-	ldrh r0, [r1]
-	cmp r0, 0
-	bne _080196D8
-	lsls r0, r5, 3
-	adds r1, r2, 0
-	adds r1, 0x38
-	adds r1, r0
-	ldr r0, [r1]
-	str r0, [r2, 0x2C]
-_080196FA:
-	pop {r4-r7}
-	pop {r0}
-	bx r0
-	thumb_func_end sub_8019660
-
         .align 2,0
