+++ src/code_801B3C0.c
@@ -501,15 +501,15 @@ void sub_801B874(void)
 
 void sub_801BB20(void)
 {
-    struct unkStruct_203B234 *preload;
-    u8 *nullText;
+    s32 loopMax;
 
-    preload = gUnknown_203B234;
-    nullText = NULL;
-    preload->unkCC[0].text = gKangaskhanTrashToolboxItem;
-    preload->unkCC[0].menuAction = 4;
-    preload->unkCC[1].text = gKangaskhanTrashReceivedItem;
-    preload->unkCC[1].menuAction = 5;
-    preload->unkCC[2].text = nullText;
-    preload->unkCC[2].menuAction = -1;
+    loopMax = 0;
+    gUnknown_203B234->unkCC[loopMax].text = gKangaskhanTrashToolboxItem;
+    gUnknown_203B234->unkCC[loopMax].menuAction = 4;
+    loopMax += 1;
+    gUnknown_203B234->unkCC[loopMax].text = gKangaskhanTrashReceivedItem;
+    gUnknown_203B234->unkCC[loopMax].menuAction = 5;
+    loopMax += 1;
+    gUnknown_203B234->unkCC[loopMax].text = NULL;
+    gUnknown_203B234->unkCC[loopMax].menuAction = -1;
 }
